ref: move parse message function inside class
